Fixed database isomorphism bug
Couldn't add C=C[O] to the database because it wasn't being seen as unique. [CH2]C=O (already in database) is a resonance structure, and isomorphism checks come back true if the resonance structures have been generated. Changed the isomorphism check to look at molecule[0] because this is unaffected by presence of other resonance structures.